# Added pagination to the aws_ec2_ami_shared table list operation and optimized the table to perform batch operation with given image IDs and owner IDs