--- a/docs/tables/aws_ec2_ami_shared.md
+++ b/docs/tables/aws_ec2_ami_shared.md
@@ -14,6 +14,7 @@ The `aws_ec2_ami_shared` table in Steampipe provides you with information about 
 **Important Notes**
 - You must specify an Owner ID or Image ID in the `where` clause (`where owner_id='`), (`where image_id='`).
 - The `aws_ec2_ami_shared` table can list any image but you must specify `owner_id` or `image_id`.
+- To optimize the query timing and API call the optional query parameter `owner_ids` or `image_ids` to perform a batch operation.
 - If you want to list all of the images in your account then you can use the `aws_ec2_ami` table.
 
 ## Examples
@@ -117,4 +118,94 @@ from
  join aws_ec2_ami_shared as ami on i.image_id = ami.image_id
 where
   ami.owner_id = '137112412989';
+```
+
+### Batch API operation, ensuring AMIs are from trusted sources
+Any AWS customer can publish an Amazon Machine Image (AMI) for other AWS customers to launch instances from. AWS only vets a handful of images in the AWS Marketplace, there is no guarantee that other publicly shared AMIs are free of vulnerabilities or malicious code. While it's common for vendors to share their software as an AMI, it's also possible someone in your organization has launched an instance from a compromised image.
+
+```sql+postgres
+with instances as (
+  select
+    instance_id,
+    instance_type,
+    account_id,
+    tags ->> 'Name' as instance_name,
+    _ctx ->> 'connection_name' as account_name,
+    instance_state,
+    region,
+    image_id
+  from
+    aws_ec2_instance
+),
+all_image_ids as (
+  select
+    json_agg(image_id)::jsonb AS image_ids  -- Cast to jsonb
+  from
+    instances
+),
+shared_ami as (
+  select
+    s.*
+  from
+    aws_ec2_ami_shared as s,
+    all_image_ids
+  where s.image_ids = all_image_ids.image_ids
+)
+select distinct
+  shared_ami.image_id as image_id,
+  shared_ami.owner_id as image_owner_id,
+  shared_ami.image_owner_alias as image_owner_name,
+  instances.instance_name,
+  instances.account_name,
+  instances.region,
+  shared_ami.name as image_name
+from
+  instances
+left join shared_ami on shared_ami.image_id=instances.image_id
+where shared_ami.image_owner_alias != 'amazon'
+and shared_ami.image_owner_alias != 'self';
+```
+
+```sql+sqlite
+with instances as (
+  select
+    instance_id,
+    instance_type,
+    account_id,
+    json_extract(tags, '$.Name') as instance_name,
+    json_extract(_ctx, '$.connection_name') as account_name,
+    instance_state,
+    region,
+    image_id
+  from
+    aws_ec2_instance
+),
+all_image_ids as (
+  select
+    json_group_array(image_id) as image_ids
+  from
+    instances
+),
+shared_ami as (
+  select
+    s.*,
+    ai.image_ids
+  from
+    aws_ec2_ami_shared as s,
+    all_image_ids as ai
+  where json_array_contains(ai.image_ids, s.image_id)
+)
+select distinct
+  shared_ami.image_id as image_id,
+  shared_ami.owner_id as image_owner_id,
+  shared_ami.image_owner_alias as image_owner_name,
+  instances.instance_name,
+  instances.account_name,
+  instances.region,
+  shared_ami.name as image_name
+from
+  instances
+left join shared_ami on shared_ami.image_id = instances.image_id
+where shared_ami.image_owner_alias != 'amazon'
+  and shared_ami.image_owner_alias != 'self';
 ```